+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 * log: log-processor collects its own cpu/heap/kafka stats to index
 * log: ActionLogContext.stat() supports adding up
 * log: ActionLogContext.track() tracks I/O reads and writes for heavy backend db, such as DB/Mongo/ES
+* web: put max forwarded ips config to prevent from x-forwarded-for clientIp spoofing
+        http().maxForwardedIPs() 
 
 ### 5.2.0 (10/28/2017 - 11/7/2017)
 * bean: removed "javax.inject:javax.inject:1", replaced with core.framework.inject, removed constructor injection support (to simplify and prepare for JDK 9)